Extract  module
