+++ executor/src/witgen/machines/block_machine.rs
@@ -5,6 +5,7 @@ use crate::witgen::column_map::ColumnMap;
 use crate::witgen::identity_processor::IdentityProcessor;
 use crate::witgen::processor::Processor;
 use crate::witgen::rows::{Row, RowFactory, RowPair, RowUpdater, UnknownStrategy};
+use crate::witgen::sequence_iterator::{IdentityInSequence, ProcessingSequenceCache, SequenceStep};
 use crate::witgen::util::try_to_simple_poly;
 use crate::witgen::{
     affine_expression::AffineResult, machines::Machine, range_constraints::RangeConstraint,
@@ -493,231 +494,5 @@ impl<'a, T: FieldElement> BlockMachine<'a, T> {
         let current = &self.data[row as usize];
         let next = &self.data[row as usize + 1];
         RowPair::new(current, next, row, fixed_data, UnknownStrategy::Unknown)
-    }
-}
-
-#[derive(Clone, Debug)]
-struct SequenceStep {
-    row_delta: i64,
-    identity: IdentityInSequence,
-}
-
-/// Goes through all rows of the block machine (plus the ones before and after)
-/// forward, backward, and forward again.
-/// In each row, iterates over all identities until no further progress is made.
-struct DefaultSequenceIterator {
-    block_size: usize,
-    identities_count: usize,
-    row_deltas: Vec<i64>,
-
-    /// Whether this is the first time the iterator is called.
-    is_first: bool,
-    /// Whether any progress was made in the current round.
-    progress_in_current_round: bool,
-    /// The current row delta index.
-    cur_row_delta_index: usize,
-    /// The current identity index.
-    cur_identity_index: usize,
-    /// The number of rounds for the current row delta.
-    /// If this number gets too large, we will assume that we're in an infinite loop and exit.
-    current_round_count: usize,
-}
-
-const MAX_ROUNDS_PER_ROW_DELTA: usize = 100;
-
-impl DefaultSequenceIterator {
-    fn new(block_size: usize, identities_count: usize) -> Self {
-        let max_row = block_size as i64 - 1;
-        DefaultSequenceIterator {
-            block_size,
-            identities_count,
-            row_deltas: (-1..=max_row)
-                .chain((-1..max_row).rev())
-                .chain(0..=max_row)
-                .collect(),
-            is_first: true,
-            progress_in_current_round: false,
-            cur_row_delta_index: 0,
-            cur_identity_index: 0,
-            current_round_count: 0,
-        }
-    }
-
-    /// Update the state of the iterator.
-    /// If we're not at the last identity in the current row, just moves to the next.
-    /// Otherwise, starts with identity 0 and moves to the next row if no progress was made.
-    fn update_state(&mut self) {
-        if !self.is_first {
-            if self.is_last_identity() {
-                self.start_next_round();
-            } else {
-                // Stay at row delta, move to next identity.
-                self.cur_identity_index += 1;
-            }
-        }
-        self.is_first = false;
-    }
-
-    fn is_last_identity(&self) -> bool {
-        let row_delta = self.row_deltas[self.cur_row_delta_index];
-
-        if row_delta + 1 == self.block_size as i64 {
-            // In the last row, we want to process one more identity, the outer query.
-            self.cur_identity_index == self.identities_count
-        } else {
-            self.cur_identity_index == self.identities_count - 1
-        }
-    }
-
-    fn start_next_round(&mut self) {
-        if !self.progress_in_current_round || self.current_round_count > MAX_ROUNDS_PER_ROW_DELTA {
-            // Move to next row delta, starting with identity 0.
-            if self.current_round_count > MAX_ROUNDS_PER_ROW_DELTA {
-                panic!("In witness generation for block machine, we have been stuck in the same row for {MAX_ROUNDS_PER_ROW_DELTA} rounds. \
-                            This is a bug in the witness generation algorithm.");
-            }
-
-            self.cur_row_delta_index += 1;
-            self.current_round_count = 0;
-        } else {
-            // Stay and current row delta, starting with identity 0.
-            self.current_round_count += 1;
-        }
-        self.cur_identity_index = 0;
-        self.progress_in_current_round = false;
-    }
-
-    pub fn report_progress(&mut self, progress_in_last_step: bool) {
-        if !self.is_first {
-            self.progress_in_current_round |= progress_in_last_step;
-        }
-    }
-
-    pub fn next(&mut self) -> Option<SequenceStep> {
-        self.update_state();
-
-        if self.cur_row_delta_index == self.row_deltas.len() {
-            // Done!
-            return None;
-        }
-
-        let row_delta = self.row_deltas[self.cur_row_delta_index];
-        let identity = if self.cur_identity_index < self.identities_count {
-            IdentityInSequence::Internal(self.cur_identity_index)
-        } else {
-            IdentityInSequence::OuterQuery
-        };
-
-        Some(SequenceStep {
-            row_delta,
-            identity,
-        })
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-enum IdentityInSequence {
-    Internal(usize),
-    OuterQuery,
-}
-
-#[derive(PartialOrd, Ord, PartialEq, Eq, Debug)]
-struct SequenceCacheKey {
-    /// For each expression on the left-hand side of the lookup, whether it is a constant.
-    known_columns: Vec<bool>,
-}
-
-impl<K, T> From<&[AffineResult<K, T>]> for SequenceCacheKey
-where
-    K: Copy + Ord,
-    T: FieldElement,
-{
-    fn from(value: &[AffineResult<K, T>]) -> Self {
-        SequenceCacheKey {
-            known_columns: value
-                .iter()
-                .map(|v| {
-                    v.as_ref()
-                        .ok()
-                        .and_then(|ex| ex.is_constant().then_some(true))
-                        .is_some()
-                })
-                .collect(),
-        }
-    }
-}
-
-enum ProcessingSequenceIterator<I: Iterator<Item = SequenceStep>> {
-    Default(DefaultSequenceIterator),
-    Cached(I),
-}
-
-impl<I: Iterator<Item = SequenceStep>> ProcessingSequenceIterator<I> {
-    fn report_progress(&mut self, progress_in_last_step: bool) {
-        match self {
-            ProcessingSequenceIterator::Default(it) => it.report_progress(progress_in_last_step),
-            ProcessingSequenceIterator::Cached(_) => {} // Progress is ignored
-        }
-    }
-}
-
-impl<I: Iterator<Item = SequenceStep>> Iterator for ProcessingSequenceIterator<I> {
-    type Item = SequenceStep;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match self {
-            ProcessingSequenceIterator::Default(it) => it.next(),
-            ProcessingSequenceIterator::Cached(it) => it.next(),
-        }
-    }
-}
-
-struct ProcessingSequenceCache {
-    block_size: usize,
-    identities_count: usize,
-    cache: BTreeMap<SequenceCacheKey, Vec<SequenceStep>>,
-}
-
-impl ProcessingSequenceCache {
-    pub fn new(block_size: usize, identities_count: usize) -> Self {
-        ProcessingSequenceCache {
-            block_size,
-            identities_count,
-            cache: Default::default(),
-        }
-    }
-
-    pub fn get_processing_sequence<K, T>(
-        &self,
-        left: &[AffineResult<K, T>],
-    ) -> ProcessingSequenceIterator<impl Iterator<Item = SequenceStep>>
-    where
-        K: Copy + Ord,
-        T: FieldElement,
-    {
-        match self.cache.get(&left.into()) {
-            Some(cached_sequence) => {
-                log::trace!("Using cached sequence");
-                ProcessingSequenceIterator::Cached(cached_sequence.clone().into_iter())
-            }
-            None => {
-                log::trace!("Using default sequence");
-                ProcessingSequenceIterator::Default(DefaultSequenceIterator::new(
-                    self.block_size,
-                    self.identities_count,
-                ))
-            }
-        }
-    }
-
-    pub fn report_processing_sequence<K, T>(
-        &mut self,
-        left: &[AffineResult<K, T>],
-        sequence: Vec<SequenceStep>,
-    ) where
-        K: Copy + Ord,
-        T: FieldElement,
-    {
-        self.cache.entry(left.into()).or_insert(sequence);
     }
 }
